4.20/21 completed, tests passing.

+++ part4/blogs/controllers/blogs.js
@@ -9,18 +9,10 @@ blogsRouter.get('', async (request, response) => {
   response.json(blogs)
 })
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
-
-blogsRouter.post('', async (request, response, next) => {
+blogsRouter.post('', async (request, response) => {
   const body = request.body
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
@@ -43,6 +35,24 @@ blogsRouter.post('', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  try {
+    // Look for the blog to delete.
+    const blogToDelete = await Blog.findById(request.params.id)
+
+    // If it can be found, check that the logged in user is the same as the creator of the blog. If not, they don't have permission to delete it.
+    if ( blogToDelete.user.toString() !== decodedToken.id.toString() ) {
+      return response.status(401).json({ error: 'invalid permissions - blog not created by user' })
+    }
+  } catch (error) {
+    // If it can't be found, return a 204 'no content' status
+    response.status(204).end()
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
